+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # Changelog
 
-## [0.2.1](https://github.com/auth0-lab/fga-js-sdk/compare/v0.2.0...v0.2.1) (2022-03-07)
+## v0.2.2
 
-feat(error-handling): expose new api error codes
+### [0.2.2](https://github.com/auth0-lab/fga-gp-sdk/compare/v0.2.1...v0.2.2) (2022-03-09)
 
-## 0.2.0 (2022-02-09)
+#### Changes
+- fix: fix for return types on 204 no content
 
+## v0.2.1
+
+### [0.2.1](https://github.com/auth0-lab/fga-go-sdk/compare/v0.2.0...v0.2.1) (2022-03-07)
+
+#### Changes
+- feat(error-handling): expose new api error codes
+
+## v0.2.0
+
+### 0.2.0 (2022-02-09)
+
+#### Changes
 - feat: update interfaces for latest api breaking changes
 - chore(deps): update dependencies